pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "microcosm-4x"
description = "A little 4X game written using Pyxel."
readme = "README.md"
requires-python = ">=3.10"
license = "GPL-3.0-only"
keywords = [
    "game",
    "4x",
    "pyxel"
]
authors = [
    { name = "ChrisNeedham24", email = "christopherneedham24@gmail.com" },
]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
# Make sure the below are kept in sync with requirements.txt.
dependencies = [
    "pyxel==1.9.0",
    "python-vlc==3.0.16120",
    "platformdirs==3.9.1",
    "pillow>=10.4.0",
    # UPnP support for networking - temporarily disabled for Windows development
    # "miniupnpc==2.0.2; platform_system != 'Windows'",
    # "miniupnpc==2.2.7; platform_system == 'Windows'"
]
version = "0.1.0"

# We use gui-scripts rather than just scripts so that Windows users can run them without a console.
[project.gui-scripts]
# This installs a 'microcosm' binary that runs the game.
microcosm = "source.game:Game"

[project.urls]
Documentation = "https://github.com/ChrisNeedham24/microcosm#readme"
Issues = "https://github.com/ChrisNeedham24/microcosm/issues"
Source = "https://github.com/ChrisNeedham24/microcosm"

[tool.hatch.build.targets.wheel]
# We don't need game_server.py because users won't be running the server, and we don't need microcosm.py because the
# gui-script defined above automatically generates an equivalent.
include = [
    "source/**",
]
exclude = [
    "source/resources/microcosm_gameplay_screenshot.png",
    "*.md",
    "source/tests/*"
]

# UV configuration
[tool.uv]
dev-dependencies = [
    "pylint==2.17.1",
    "coverage==6.5.0",
    "nuitka==1.5.5",
    "imageio==2.26.1",
    "hatch==1.14.0",
]
